amdgcn: Fix various unrecognized pattern issues with add<mode>3_vcc_dup
The patterns did not accept inline immediate constants, even though the
hardware instructions do, which has lead to some errors in some patches I'm
working on.

Also the VCC update RTL was using the wrong operands in the wrong places.  This
appears to have been harmless(?) but is definitely not intended.

gcc/ChangeLog:

	* config/gcn/gcn-valu.md (add<mode>3_vcc_dup<exec_vcc>): Change
	operand 2 to allow gcn_alu_operand.  Swap the operands in the VCC
	update RTL.
	(add<mode>3_vcc_zext_dup): Likewise.
	(add<mode>3_vcc_zext_dup_exec): Likewise.
	(add<mode>3_vcc_zext_dup2): Likewise.
	(add<mode>3_vcc_zext_dup2_exec): Likewise.