Make the RTL frontend set REG_NREGS correctly
While working on a new testcase that uses the RTL frontend,
I hit a bug where a (reg ...) that spans multiple hard registers
had REG_NREGS set to 1.  This caused various things to misbehave.
For example, if the (reg ...) in question was used as crtl->return_rtx,
only the first register in the group would be marked as live on exit.

gcc/
	* read-rtl-function.cc (function_reader::read_rtx_operand_r): Use
	hard_regno_nregs to work out REG_NREGS for hard registers.