libstdc++: Implement std::chrono::current_zone() for Windows [PR108409]
On Windows there is no API to get the current time zone as IANA name,
instead Windows has its own zones. But there exists a mapping provided
by the Unicode Consortium. This patch adds a script to convert the XML
file with the mapping to a lookup table and adds a Windows code path to
use that mapping.

libstdc++-v3/Changelog:

	Implement std::chrono::current_zone() for Windows

	PR libstdc++/108409
	* scripts/gen_windows_zones_map.py: New file, generates
	windows_zones-map.h.
	* src/c++20/windows_zones-map.h: New file, contains the look up
	table.
	* src/c++20/tzdb.cc (tzdb::current_zone): Add Windows code path.

Signed-off-by: Björn Schäpers <bjoern@hazardy.de>