Avoid vect_is_simple_use call from get_load_store_type
This isn't the required refactoring of vect_check_gather_scatter
but it avoids a now unnecessary call to vect_is_simple_use which
is problematic because it looks at STMT_VINFO_VECTYPE which we
want to get rid of.  SLP build already ensures vect_is_simple_use
on all lane defs, so all we need is to populate the offset_vectype
and offset_dt which is not always set by vect_check_gather_scatter.
That's both easy to get from the SLP child directly.

	* tree-vect-stmts.cc (get_load_store_type): Do not use
	vect_is_simple_use to fill gather/scatter offset operand
	vectype and dt.