Avoid vect_is_simple_use call from vectorizable_reduction
When analyzing the reduction cycle we look to determine the
reduction input vector type, for lane-reducing ops we look
at the input but instead of using vect_is_simple_use which
is problematic for SLP we should simply get at the SLP
operands vector type.  If that's not set and we make up one
we should also ensure it stays so.

	* tree-vect-loop.cc (vectorizable_reduction): Avoid
	vect_is_simple_use and record a vector type if we come
	up with one.