Reject single lane vector types for SLP build
The following makes us never consider vector(1) T types for
vectorization and ensures this during SLP build.  This is a
long-standing issue for BB vectorization and when we remove
early loop vector type setting we lose the single place we have
that rejects this for loops.

Once we implement partial loop vectorization we should revisit
this, but then use the original scalar types for the unvectorized
parts.

	* tree-vect-slp.cc (vect_build_slp_tree_1): Reject
	single-lane vector types.

	* gcc.dg/vect/bb-slp-39.c: Adjust.