properly compute fp/mode for scalar ops for vectorizer costing
The x86 add_stmt_hook relies on the passed vectype to determine
the mode and whether it is FP for a scalar operation.  This is
unreliable now for stmts involving patterns and in the future when
there is no vector type passed for scalar operations.

To be least disruptive I've kept using the vector type if it is passed.

	* config/i386/i386.cc (ix86_vector_costs::add_stmt_cost): Use
	the LHS of a scalar stmt to determine mode and whether it is FP.