aarch64: Guard VF-based costing with !m_costing_for_scalar
g:4b47acfe2b626d1276e229a0cf165e934813df6c caused a segfault
in aarch64_vector_costs::analyze_loop_vinfo when costing scalar
code, since we'd end up dividing by a zero VF.

Much of the structure of the aarch64 costing code dates from
a stage 4 patch, when we had to work within the bounds of what
the target-independent code did.  Some of it could do with a
rework now that we're not so constrained.

This patch is therefore an emergency fix rather than the best
long-term solution.  I'll revisit when I have more time to think
about it.

gcc/
	* config/aarch64/aarch64.cc (aarch64_vector_costs::add_stmt_cost):
	Guard VF-based costing with !m_costing_for_scalar.