libstdc++: Use basic_string_view for std::bitset deduced types
This change was part of by P2697R1 (Interfacing bitset with string_view)
and should be slightly cheaper to instantiate.

We should consider using basic_string_view for C++17, C++20, and C++23
as well. This patch just conservatively changes it for C++26 to match
the working draft. It's conceivable that a program-defined
specialization of basic_string<_CharT> or basic_string_view<_CharT> will
observe a difference and be affected by this change.

libstdc++-v3/ChangeLog:

	* include/std/bitset (__bitset::__string) [__cpp_lib_bitset]:
	Change alias to refer to basic_string_view instead.

Reviewed-by: Tomasz Kamiński <tkaminsk@redhat.com>