Fix some auto-profile issues
This patch fixes minor things that has cumulated in my tree.  Except for
formating fixes an important change is that seen set is now kept up to date.
Oriignal code first populated it for all string in the string table but now
gimple matching may introduce new ones that needs to be checked for match with
symbol table as well.

This makes imagemagic of spec2017 to be faster with auto-fdo then without at
least when trained with ref run.  Train run has problem since it does not train
the innermost loop at all, so even with normal PGO it is slower then without.

autorpfoiledbootstrapped/regtested x86_64-linux, comitted.

gcc/ChangeLog:

	* auto-profile.cc (function_instance::~function_instance):
	Move down in source.
	(string_table::get_cgraph_node): New member function with
	logic broken out from ...
	(function_instance::get_cgraph_node): ... here.
	(match_with_target): Fix formating.
	(function_instance::match): Fix formating; do not use iterators
	after modifying map; remove incorrect set of warned flag.
	(autofdo_source_profile::offline_external_functions): Keep
	seen set up to date.
	(function_instance::read_function_instance): Fix formating.