fortran: Generate array reallocation out of loops
Generate the array reallocation on assignment code before entering the
scalarization loops.  This doesn't move the generated code itself,
which was already put before the outermost loop, but only changes the
current scope at the time the code is generated.  This is a prerequisite
for a followup patch that makes the reallocation code create new
variables.  Without this change the new variables would be declared in
the innermost loop body and couldn't be used outside of it.

gcc/fortran/ChangeLog:

	* trans-expr.cc (gfc_trans_assignment_1): Generate array
	reallocation code before entering the scalarisation loops.