libstdc++: Add comments to __unicode::_Utf_iterator
Add comments documenting what it does and how it does it.

Also reorder the if-else in operator++ so that we check whether to
iterate over code units in the local buffer before checking whether to
refill that buffer. That seems the more natural way to structure the
function.

libstdc++-v3/ChangeLog:

	* include/bits/unicode.h (__unicode::_Utf_iterator): Add
	comments.
	(__unicode:_Utf_iterator::operator++()): Check whether to
	iterate over the buffer first.

Reviewed-by: Tomasz Kamiński <tkaminsk@redhat.com>